+++ src-tauri/Cargo.toml
@@ -16,6 +16,7 @@ tauri-build = { version = "1.0.0-beta.2" }
 serde_json = "1.0"
 serde = { version = "1.0", features = [ "derive" ] }
 fs2 = "0.4.3"
+
 [target.'cfg(windows)'.dependencies]
 winreg = "0.10.1"
