+++ src-tauri/src/main.rs
@@ -2,27 +2,29 @@
   all(not(debug_assertions), target_os = "windows"),
   windows_subsystem = "windows"
 )]
-use std::path::PathBuf;
+
+#[cfg(target_os = "windows")]
+mod windows;
 
 use serde::Serialize;
-
+use std::path::PathBuf;
 use tauri::{
   api, CustomMenuItem, Manager, SystemTray, SystemTrayEvent, SystemTrayMenu, SystemTrayMenuItem,
   WindowEvent,
 };
 
 #[derive(Serialize)]
-struct S {
+struct DiskStats {
   free_bytes: u64,
   total_bytes: u64,
 }
 
 #[tauri::command]
-fn get_disk_stats(dir: String) -> S {
+fn get_disk_stats(dir: String) -> DiskStats {
   println!("{}", dir.to_string());
   let free: u64 = fs2::available_space(&dir).expect("error");
   let total: u64 = fs2::total_space(&dir).expect("error");
-  return S {
+  return DiskStats {
     free_bytes: free,
     total_bytes: total,
   };
@@ -32,61 +34,6 @@ fn get_disk_stats(dir: String) -> S {
 fn get_this_binary() -> PathBuf {
   let bin = api::process::current_binary();
   return bin.unwrap();
-}
-#[cfg(target_os = "windows")]
-
-//TODO Refactor into special case functions instead of general utility functions
-mod win_functions {
-  #[tauri::command]
-  pub fn winreg_get(sub_key: String, value: String) -> String {
-    let hkcu = winreg::RegKey::predef(winreg::enums::HKEY_CURRENT_USER);
-    let subkey = hkcu.open_subkey(sub_key).unwrap();
-    let val = subkey.get_value(value);
-    let result: String;
-    match val {
-      Ok(content) => {
-        println!("File content: {}", content);
-        result = content;
-      }
-      Err(error) => {
-        eprintln!("There was an error: {}", error);
-        result = error.to_string();
-      }
-    }
-    return result;
-  }
-  #[tauri::command]
-  pub fn winreg_set(sub_key: String, set_key: String, value: String) -> String {
-    let hkcu = winreg::RegKey::predef(winreg::enums::HKEY_CURRENT_USER);
-    // let subkey = hkcu.open_subkey(sub_key).unwrap();
-    let (subkey, _disp) = hkcu.create_subkey(&sub_key).unwrap();
-
-    let val = subkey.set_value(set_key, &value);
-    let mut result: String = "success".to_string();
-    match val {
-      Ok(()) => {}
-      Err(error) => {
-        eprintln!("There was an error: {}", error);
-        result = error.to_string();
-      }
-    }
-    return result;
-  }
-  #[tauri::command]
-  pub fn winreg_delete(sub_key: String, set_key: String) -> String {
-    let hkcu = winreg::RegKey::predef(winreg::enums::HKEY_CURRENT_USER);
-    let (subkey, _disp) = hkcu.create_subkey(&sub_key).unwrap();
-    let val = subkey.delete_value(set_key);
-    let mut result: String = "success".to_string();
-    match val {
-      Ok(()) => {}
-      Err(error) => {
-        eprintln!("There was an error: {}", error);
-        result = error.to_string();
-      }
-    }
-    return result;
-  }
 }
 
 fn main() {
@@ -98,7 +45,7 @@ fn main() {
     .add_native_item(SystemTrayMenuItem::Separator)
     .add_item(quit);
   let tray = SystemTray::new().with_menu(tray_menu);
-  tauri::Builder::default()
+  let builder = tauri::Builder::default()
     .setup(|app| {
       let window = app.get_window("main").unwrap();
 
@@ -146,18 +93,16 @@ fn main() {
       }
       _ => {}
     })
-    .invoke_handler(
-      #[cfg(target_os = "windows")]
-      tauri::generate_handler![
-        get_disk_stats,
-        get_this_binary,
-        win_functions::winreg_get,
-        win_functions::winreg_set,
-        win_functions::winreg_delete,
-      ],
-      #[cfg(not(target_os = "windows"))]
-      tauri::generate_handler![get_disk_stats, get_this_binary],
-    )
+    .invoke_handler(tauri::generate_handler![get_disk_stats, get_this_binary]);
+
+  #[cfg(target_os = "windows")]
+  let builder = builder.invoke_handler(tauri::generate_handler![
+    windows::winreg_get,
+    windows::winreg_set,
+    windows::winreg_delete,
+  ]);
+
+  builder
     .run(tauri::generate_context!())
     .expect("error while running tauri application");
 }
